# Micro-optimize double comparison